feat: :art: I add the first commit to the app, with integration of environment variables, the entity “book”, exception filters,  swageer configuration and connection with database in mongo db.
I add the first commit to the app, with integration of environment variables, the entity “book”, exception filters,  swageer configuration and connection with database in mongo db.

I add the first commit to the app, with integration of environment variables, the entity “book”, exception filters,  swageer configuration and connection with database in mongo db.